+++ google-cloud/container-linux/kubernetes/workers/variables.tf
@@ -20,9 +20,9 @@ variable "count" {
   description = "Number of worker compute instances the instance group should manage"
 }
 
-variable "zone" {
+variable "region" {
   type        = "string"
-  description = "Google zone that compute instances in the group should be created in (e.g. gcloud compute zones list)"
+  description = "Google Cloud region to create a regional managed group of workers (e.g. us-central1, see `gcloud compute regions list`)."
 }
 
 variable "machine_type" {